+++ src/ruby/app/models/keyboard.rb
@@ -454,7 +454,7 @@ class Keyboard
     @anchor_left = true # so-called "master left"
     @uart_pin = 1
     $rgb = nil
-    $encoders = Array.new
+    @encoders = Array.new
     @partner_encoders = Array.new
     @macro_keycodes = Array.new
     @buffer = Buffer.new("picoirb")
@@ -477,12 +477,12 @@ class Keyboard
         if @anchor_left
           if @anchor == feature.left? # XNOR
             feature.init_pins
-            $encoders << feature
+            @encoders << feature
           end
         else
           if @anchor != feature.left? #XOR
             feature.init_pins
-            $encoders << feature
+            @encoders << feature
           end
         end
         if @anchor && (@anchor_left != feature.left?)
@@ -490,7 +490,7 @@ class Keyboard
         end
       else
         feature.init_pins
-        $encoders << feature
+        @encoders << feature
       end
     end
   end
@@ -755,9 +755,9 @@ class Keyboard
       end
     end
     report_hid(modifier, "#{c}\000\000\000\000\000")
-    sleep_ms 10
+    sleep_ms 1
     report_hid(0, "\000\000\000\000\000\000")
-    sleep_ms 10
+    sleep_ms 1
   end
 
   # **************************************************************
@@ -935,7 +935,7 @@ class Keyboard
           block.call
         end
 
-        $encoders.each do |encoder|
+        @encoders.each do |encoder|
           encoder.consume_rotation_anchor
         end
 
@@ -971,7 +971,7 @@ class Keyboard
         end
       else
         # Partner
-        $encoders.each do |encoder|
+        @encoders.each do |encoder|
           data = encoder.consume_rotation_partner
           uart_partner_push8(data) if data && data > 0
         end