+++ src/ruby/app/models/rotary_encoder.rb
@@ -3,7 +3,6 @@ class RotaryEncoder
     puts "Initializing RotaryEncoder ..."
     @pin_a = pin_a
     @pin_b = pin_b
-    @rotation = 0
     @partner_keycode_cw = 0
     @partner_keycode_ccw = 0
     @split_left = true
@@ -12,7 +11,7 @@ class RotaryEncoder
   end
 
   def init_pins
-    init_encoder(@pin_a, @pin_b)
+    @number = init_encoder(@pin_a, @pin_b)
   end
 
   def create_keycodes(encoders_size)
@@ -40,10 +39,6 @@ class RotaryEncoder
 
   attr_accessor :split_left
 
-  def read(index)
-    @rotation += read_encoder(index, @pin_a, @pin_b)
-  end
-
   def clockwise(&block)
     @proc_cw = block
   end
@@ -68,25 +63,22 @@ class RotaryEncoder
   end
 
   def consume_rotation_anchor
-    # ignore a value between -1..1
-    if @rotation > 1
+    case consume_encoder(@number)
+    when 1
       @proc_cw.call
-    elsif @rotation < -1
+    when -1
       @proc_ccw.call
     end
-    @rotation = 0
   end
 
   def consume_rotation_partner
-    # ignore values of 1 and -1
-    res = if @rotation > 1 && @proc_cw
+    case consume_encoder(@number)
+    when 1
       @partner_keycode_cw
-    elsif @rotation < -1 && @proc_ccw
+    when -1
       @partner_keycode_ccw
     else
       0
     end
-    @rotation = 0
-    return res
   end
 end