+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change log
 
-## 0.9.8 on 2021/12/xx
+## 0.9.8 on 2021/12/10
 ### RotaryEncoder enhancement
 - Multiple encoders can be configured on a unit🎛
-- `:RGB_xxx` keycodes work with an encoder🌈
+- `:RGB_xxx` keycodes work with encoders🌈
+- The accuracy of ratation made a big improvement😎
 
 ## 0.9.7 on 2021/12/02
 ### Bug fix